Add files via upload
For my this project, I wrote a C++ program that calculates and displays the charges for a patient’s hospital stay and writes the report to a file named “hospital.txt”. The program first asks whether the patient was admitted as an inpatient or outpatient. If the patient is an inpatient, I prompt for the number of days spent in the hospital, the daily rate, hospital service charges, and medication charges. If the patient is an outpatient, I only ask for the service and medication charges. I divided the program into functions to make it organized and reusable. I also created a validation function that ensures no input is less than zero, prompting the user to reenter any invalid data. To calculate the total charges, I implemented two overloaded functions: one for inpatient data and another for outpatient data. Each function calculates and returns the total charges, which are then displayed on the screen and written to “hospital.txt”. This project helped me practice function overloading, input validation, file handling, and modular programming in C++.